vouge/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.1
sqlparse==0.5.5
whitenoise==6.12.0
